Added custom exception, exception handling

+++ src/main/java/com/musicians/d2ovj9haladojavalev/controller/ArtistController.java
@@ -4,6 +4,7 @@ import com.musicians.d2ovj9haladojavalev.dto.ArtistDTO;
 import com.musicians.d2ovj9haladojavalev.entity.Album;
 import com.musicians.d2ovj9haladojavalev.entity.Artist;
 import com.musicians.d2ovj9haladojavalev.entity.Musician;
+import com.musicians.d2ovj9haladojavalev.exception.DataNotFoundException;
 import com.musicians.d2ovj9haladojavalev.service.AlbumService;
 import com.musicians.d2ovj9haladojavalev.service.ArtistService;
 import com.musicians.d2ovj9haladojavalev.service.MusicianService;
@@ -24,7 +25,8 @@ public class ArtistController {
     private final AlbumService albumService;
 
     @Autowired
-    public ArtistController(ArtistService theArtistService, MusicianService musicianService, AlbumService albumService) {
+    public ArtistController(ArtistService theArtistService,
+                            MusicianService musicianService, AlbumService albumService) {
         artistService = theArtistService;
         this.musicianService = musicianService;
         this.albumService = albumService;
@@ -32,29 +34,49 @@ public class ArtistController {
 
     @GetMapping("/artist/{id}")
     public Artist getArtist(@PathVariable Long id) {
-        return artistService.getArtist(id);
+        Artist artist = artistService.getArtist(id);
+        if (artist == null) {
+            throw new DataNotFoundException("...Whoops! No data found.");
+        }
+        return artist;
     }
 
     // ehelyett DTO
     @GetMapping("/artist")
     public List<Artist> getAllArtist() {
-        return artistService.getAllArtist();
+        List<Artist> artistList = artistService.getAllArtist();
+        if (artistList.isEmpty()) {
+            throw new DataNotFoundException("...Whoops! No data found.");
+        }
+        return artistList;
     }
 
     // ez lesz a dto-s impl
     @GetMapping("/artistdto")
     public List<ArtistDTO> getAllArtistDto() {
-        return artistService.getAllArtistDto();
+        List<ArtistDTO> artistList = artistService.getAllArtistDto();
+        if (artistList.isEmpty()) {
+            throw new DataNotFoundException("...Whoops! No data found.");
+        }
+        return artistList;
     }
 
     @GetMapping("/artist/genre/{genre}")
     public ArrayList<Artist> getAllArtistByGenre(@PathVariable String genre) {
-        return artistService.getAllArtistByGenre(genre);
+        ArrayList<Artist> artists = artistService.getAllArtistByGenre(genre);
+        if (artists.isEmpty()) {
+            throw new DataNotFoundException("...Whoops! No data found.");
+        }
+        return artists;
     }
 
     @GetMapping("artist/{year}")
     public ArrayList<Artist> getAllArtistByYear(@PathVariable int year) {
-        return artistService.getAllArtistByYear(year);
+        ArrayList<Artist> artists = artistService.getAllArtistByYear(year);
+        if (artists.isEmpty()) {
+            throw new DataNotFoundException("...Whoops! No data found.");
+        }
+        return artists;
     }
 
     @PostMapping("/artist")
@@ -93,6 +115,9 @@ public class ArtistController {
     ) {
         Artist artist = artistService.getArtist(artistId);
         Album album = albumService.getAlbumById(albumId);
+        if (artist == null || album == null) {
+            throw new DataNotFoundException("...Whoops! No data found.");
+        }
         artist.addAlbum(album);
         album.setArtist(artist);
         artistService.addArtist(artist);
@@ -100,8 +125,14 @@ public class ArtistController {
     }
 
     @DeleteMapping("/artist/{id}")
-    public void deleteArtist(@PathVariable Long id) {
-        artistService.deleteArtist(id);
+    public String deleteArtist(@PathVariable Long id) {
+        Artist artist = artistService.getArtist(id);
+        if (artist == null) {
+            throw new DataNotFoundException("...Whoops! Artist with id " + id + " not found.");
+        } else {
+            artistService.deleteArtist(id);
+        }
+        return "Artist deleted with id: " + id;
     }
